YYS_GameHelper/requirements.txt
# 核心依赖
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
PyYAML>=5.4.0
colorlog>=6.0.0

# ADB连接
pure-python-adb>=0.3.0

# OCR功能 (可选)
# paddlepaddle>=2.0.0
# paddleocr>=2.0.0

# 图像处理增强 (可选)
# scikit-image>=0.18.0
# matplotlib>=3.3.0

# 打包工具
nuitka>=1.8.0